Break out askToken dialog
So it can be reused in the link flow.
